innovfarma/requirements.txt
Flask>=2.2
Flask-SQLAlchemy>=3.0
Flask-Migrate>=4.0
Flask-Login>=0.6
python-dotenv>=1.0
Flask-Cors>=3.0
mysqlclient>=2.1
Flask-PyMongo>=2.3
pymongo>=4.3
